requirements.txt
Django==2.2.2
djangorestframework==3.9.4
pip==19.1.1
pylint==2.3.1